tests/wpt/web-platform-tests/css/css-break/parsing/break-inside-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fragmentation Module Level 3: getComputedStyle().breakAfter</title>
<link rel="help" href="https://drafts.csswg.org/css-break/#propdef-break-inside">
<meta name="assert" content="break-inside computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("break-inside", "auto");
test_computed_value("break-inside", "avoid");
test_computed_value("break-inside", "avoid-page");
test_computed_value("break-inside", "avoid-column");
test_computed_value("break-inside", "avoid-region");
</script>
</body>
</html>
