src/third_party/blink/web_tests/external/wpt/css/css-break/parsing/break-inside-computed-expected.txt

This is a testharness.js-based test.
PASS Property break-inside value 'auto' computes to 'auto'
PASS Property break-inside value 'avoid' computes to 'avoid'
PASS Property break-inside value 'avoid-page' computes to 'avoid-page'
PASS Property break-inside value 'avoid-column' computes to 'avoid-column'
FAIL Property break-inside value 'avoid-region' computes to 'avoid-region' assert_equals: expected "avoid-region" but got "auto"
Harness: the test ran to completion.
